@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "jasmine-node spec",
-    "start": "gulp watch"
+    "start": "webpack-dev-server --progress --colors --content-base dist/",
+    "build": "webpack -p"
   },
   "repository": {
     "type": "git",
@@ -26,15 +27,22 @@
   },
   "homepage": "https://github.com/banderson/reactive-stock-ticker-demo",
   "devDependencies": {
+    "babel-core": "^4.7.16",
+    "babel-loader": "^4.2.0",
+    "del": "^0.1.3",
+    "flux": "^2.0.1",
     "gulp": "^3.8.10",
     "gulp-browserify": "^0.5.0",
     "gulp-concat": "^2.2.0",
-    "react": "^0.11.1",
-    "reactify": "^0.14.0",
-    "flux": "^2.0.1",
-    "del": "^0.1.3",
     "gulp-livereload": "^2.1.1",
+    "react": "^0.13.0",
     "react-tools": "~0.11.2",
-    "vinyl-source-stream": "^1.0.0"
+    "reactify": "^0.14.0",
+    "vinyl-source-stream": "^1.0.0",
+    "webpack": "^1.7.3"
+  },
+  "dependencies": {
+    "jsx-loader": "^0.12.2",
+    "style-loader": "^0.8.3"
   }
 }
