@@ -1,56 +1,26 @@
-/*** @jsx React.DOM ***/
-"use strict";
-var React = require('react');
-var DataStore = require('../stores/DataStore');
-var ActionCreator = require('../actions/DataActionCreators');
+import React from 'react';
 
-var Ticker = React.createClass({
+let {PropTypes} = React;
 
-  getInitialState: function(){
+export default React.createClass({
+
+  getDefaultProps: function(){
     return {
-      trx: [],
-      symbols: ['GOOG', 'AAPL', 'FB'],
+      transactions: [],
       running: false
     };
   },
 
-  componentDidMount: function() {
-    this._toggle();
-    DataStore.addChangeListener(this._onChange);
-  },
-
-  componentWillUnmount: function() {
-    DataStore.removeChangeListener(this._onChange);
-  },
-
-  _onChange: function() {
-    this.setState({trx: DataStore.getAll()});
-  },
-
-  _toggle: function() {
-    if (this.state.running) {
-      this._stopTicker();
-    } else {
-      this._startTicker();
-    }
-  },
-
-  _startTicker: function() {
-    ActionCreator.startTicker(this.state.symbols);
-    this.setState({running: true});
-    setTimeout(this._stopTicker, 10000);
-  },
-
-  _stopTicker: function() {
-    ActionCreator.stopTicker(this.state.symbols);
-    this.setState({running: false});
-  },
-
-  _clear: function() {
-    this.setState({trx: []});
+  propTypes: {
+    transactions: PropTypes.array.isRequired,
+    running: PropTypes.bool.isRequired,
+    onStart: PropTypes.func.isRequired,
+    onStop: PropTypes.func.isRequired,
+    onReset: PropTypes.func.isRequired,
   },
 
   render: function() {
+    let {transactions, running, onReset} = this.props;
     return (
       <div>
         <table className="table table-bordered table-condensed">
@@ -58,11 +28,11 @@ var Ticker = React.createClass({
             <tr>
               <td colSpan="2">
                 <button className="btn btn-primary" onClick={this._toggle}>
-                  {this.state.running ? 'Stop' : 'Start/Resume'}
+                  {running ? 'Stop' : 'Start/Resume'}
                 </button>
               </td>
               <td colSpan="2">
-                <button onClick={this._clear} className="btn btn-success">
+                <button onClick={onReset} className="btn btn-success">
                   <i className="glyphicon glyphicon-refresh" /> Reset
                 </button>
               </td>
@@ -75,21 +45,22 @@ var Ticker = React.createClass({
             </tr>
           </thead>
           <tbody>
-            {this.state.trx.map(function(trx) {
-              return (
-                <tr key={trx.key} className={trx.category}>
-                  <td>{trx.time}</td>
-                  <td>{trx.ticker}</td>
-                  <td>{trx.price}</td>
-                  <td>{trx.perc}</td>
-                </tr>
-                );
-            })}
+            {transactions.map(trx =>
+              <tr key={trx.key} className={trx.category}>
+                <td>{trx.time}</td>
+                <td>{trx.ticker}</td>
+                <td>{trx.price}</td>
+                <td>{trx.perc}</td>
+              </tr>
+            )}
           </tbody>
         </table>
       </div>
     );
+  },
+
+  _toggle(e) {
+    let fn = (this.props.running) ? 'onStop' : 'onStart';
+    this.props[fn](e);
   }
 });
-
-module.exports = Ticker;
