@@ -1,12 +1,10 @@
-/*** @jsx React.DOM */
-
-var React = require('react'),
-    Ticker = require('./components/ticker.jsx'),
-    MutationReporter = require('./components/mutations.jsx');
+import React from 'react';
+import App from './components/app.jsx';
 
 var scoreboard = document.getElementById('scoreboard'),
-    main = document.getElementById('main'),
-    mutationReporter = MutationReporter({target: main});
+    main = document.getElementById('main');
+    // mutationReporter = MutationReporter({target: main});
 
-React.renderComponent(<Ticker />, main);
-React.renderComponent(<MutationReporter target={main} />, scoreboard);
+React.render(<App />, main);
+// React.renderComponent(MutationReporter({recording: "false", target: main}), scoreboard);
+// React.renderComponent(<Results data={results} seconds={} />, results);
